+++ Changelog.md
@@ -4,7 +4,7 @@ Changelog
 Version 0.2.0
 -------------
 * added sprite includes for :sprite entries
-* added support for UML use case and state views
+* added support for UML use case, state and class views/diagrams
 * enhanced documentation
 
 Version 0.1.0